splitting astroport mod
