+++ e2e/examples/eth_vault/strategist/mod.rs
@@ -1,2 +1,3 @@
+mod astroport;
 mod bridge;
 pub mod client;
